[17/x] Move context-specific methods into context fields code to discourage misuse
