+++ packages/eas-cli/src/commandUtils/EasCommand.ts
@@ -1,10 +1,4 @@
-import { EasJsonAccessor, EasJsonUtils } from '@expo/eas-json';
-import * as PackageManagerUtils from '@expo/package-manager';
 import { Command } from '@oclif/core';
-import chalk from 'chalk';
-import fs from 'fs-extra';
-import path from 'path';
-import semver from 'semver';
 
 import {
   AnalyticsEvent,
@@ -12,120 +6,11 @@ import {
   initAsync as initAnalyticsAsync,
   logEvent,
 } from '../analytics/rudderstackClient';
-import { learnMore } from '../log';
-import { getExpoConfig } from '../project/expoConfig';
-import { findProjectRootAsync, getProjectIdAsync } from '../project/projectUtils';
-import { Actor, getUserAsync } from '../user/User';
-import { ensureLoggedInAsync } from '../user/actions';
-import { easCliVersion } from '../utils/easCli';
-import { setVcsClient } from '../vcs';
-import GitClient from '../vcs/clients/git';
-
-async function applyCliConfigAsync(projectDir: string): Promise<void> {
-  const easJsonAccessor = new EasJsonAccessor(projectDir);
-  const config = await EasJsonUtils.getCliConfigAsync(easJsonAccessor);
-  if (config?.version && !semver.satisfies(easCliVersion, config.version)) {
-    throw new Error(
-      `You are on eas-cli@${easCliVersion} which does not satisfy the CLI version constraint in eas.json (${config.version})`
-    );
-  }
-  if (config?.requireCommit) {
-    setVcsClient(new GitClient());
-  }
-}
-
-async function ensureEasCliIsNotInDependenciesAsync(projectDir: string): Promise<void> {
-  let printCliVersionWarning = false;
-
-  const consoleWarn = (msg?: string): void => {
-    if (msg) {
-      // eslint-disable-next-line no-console
-      console.warn(chalk.yellow(msg));
-    } else {
-      // eslint-disable-next-line no-console
-      console.warn();
-    }
-  };
-
-  if (await isEasCliInDependenciesAsync(projectDir)) {
-    printCliVersionWarning = true;
-    consoleWarn(`Found ${chalk.bold('eas-cli')} in your project dependencies.`);
-  }
-
-  const maybeRepoRoot = PackageManagerUtils.findWorkspaceRoot(projectDir) ?? projectDir;
-  if (maybeRepoRoot !== projectDir && (await isEasCliInDependenciesAsync(maybeRepoRoot))) {
-    printCliVersionWarning = true;
-    consoleWarn(`Found ${chalk.bold('eas-cli')} in your monorepo dependencies.`);
-  }
-
-  if (printCliVersionWarning) {
-    consoleWarn(
-      `It's recommended to use the ${chalk.bold(
-        '"cli.version"'
-      )} field in eas.json to enforce the ${chalk.bold('eas-cli')} version for your project.`
-    );
-    consoleWarn(
-      learnMore('https://github.com/expo/eas-cli#enforcing-eas-cli-version-for-your-project')
-    );
-    consoleWarn();
-  }
-}
-
-async function isEasCliInDependenciesAsync(dir: string): Promise<boolean> {
-  const packageJsonPath = path.join(dir, 'package.json');
-  const packageJson = JSON.parse(await fs.readFile(packageJsonPath, 'utf8'));
-  return (
-    packageJson?.dependencies?.['eas-cli'] !== undefined ||
-    packageJson?.devDependencies?.['eas-cli'] !== undefined
-  );
-}
-
-async function findProjectDirAndVerifyProjectSetupAsync(): Promise<string> {
-  const projectDir = await findProjectRootAsync();
-  await applyCliConfigAsync(projectDir);
-  await ensureEasCliIsNotInDependenciesAsync(projectDir);
-  return projectDir;
-}
-
-export interface ContextOptions {
-  nonInteractive: boolean;
-}
-
-export abstract class ContextField<T> {
-  abstract getValueAsync(options: ContextOptions): Promise<T>;
-}
-
-export class ProjectDirContextField extends ContextField<string> {
-  async getValueAsync(): Promise<string> {
-    return await findProjectDirAndVerifyProjectSetupAsync();
-  }
-}
-
-export class ProjectIdContextField extends ContextField<string> {
-  async getValueAsync({ nonInteractive }: ContextOptions): Promise<string> {
-    const projectDir = await findProjectDirAndVerifyProjectSetupAsync();
-    const exp = getExpoConfig(projectDir);
-    return await getProjectIdAsync(exp, { nonInteractive });
-  }
-}
-
-export class OptionalProjectIdContextField extends ContextField<string | undefined> {
-  async getValueAsync({ nonInteractive }: ContextOptions): Promise<string | undefined> {
-    const projectDir = await findProjectDirAndVerifyProjectSetupAsync();
-    if (!projectDir) {
-      return undefined;
-    }
-
-    const exp = getExpoConfig(projectDir);
-    return await getProjectIdAsync(exp, { nonInteractive });
-  }
-}
-
-export class ActorContextField extends ContextField<Actor> {
-  async getValueAsync({ nonInteractive }: ContextOptions): Promise<Actor> {
-    return await ensureLoggedInAsync({ nonInteractive });
-  }
-}
+import { getUserAsync } from '../user/User';
+import ActorContextField from './context/ActorContextField';
+import ContextField from './context/ContextField';
+import ProjectDirContextField from './context/ProjectDirContextField';
+import ProjectIdContextField from './context/ProjectIdContextField';
 
 export const EASCommandProjectDirContext = {
   projectDir: new ProjectDirContextField(),
@@ -133,10 +18,6 @@ export const EASCommandProjectDirContext = {
 
 export const EASCommandProjectIdContext = {
   projectId: new ProjectIdContextField(),
-};
-
-export const EASCommandProjectIdIfProjectDirContext = {
-  projectId: new OptionalProjectIdContextField(),
 };
 
 export const EASCommandLoggedInContext = {
