+++ packages/eas-cli/src/commandUtils/context/ContextField.ts
@@ -1,0 +1,7 @@
+export interface ContextOptions {
+  nonInteractive: boolean;
+}
+
+export default abstract class ContextField<T> {
+  abstract getValueAsync(options: ContextOptions): Promise<T>;
+}
