+++ packages/eas-cli/src/commands/account/login.ts
@@ -1,13 +1,13 @@
 import EasCommand from '../../commandUtils/EasCommand';
+import ActorContextField from '../../commandUtils/context/ActorContextField';
 import Log from '../../log';
-import { showLoginPromptAsync } from '../../user/actions';
 
 export default class AccountLogin extends EasCommand {
   static override description = 'log in with your Expo account';
   static override aliases = ['login'];
 
   async runAsync(): Promise<void> {
-    await showLoginPromptAsync();
+    await ActorContextField['showLoginPromptAsync']();
     Log.log('Logged in');
   }
 }
