+++ packages/eas-cli/src/commands/project/__tests__/init.test.ts
@@ -2,17 +2,17 @@ import { AppJSONConfig, PackageJSONConfig, getConfig } from '@expo/config';
 import chalk from 'chalk';
 import { vol } from 'memfs';
 
+import ActorContextField from '../../../commandUtils/context/ActorContextField';
+import ProjectDirContextField from '../../../commandUtils/context/ProjectDirContextField';
+import ProjectIdContextField from '../../../commandUtils/context/ProjectIdContextField';
 import { jester } from '../../../credentials/__tests__/fixtures-constants';
 import { AppMutation } from '../../../graphql/mutations/AppMutation';
 import { findProjectIdByAccountNameAndSlugNullableAsync } from '../../../project/fetchOrCreateProjectIDForWriteToConfigWithConfirmationAsync';
-import { findProjectRootAsync, saveProjectIdToAppConfigAsync } from '../../../project/projectUtils';
 import { confirmAsync, promptAsync } from '../../../prompts';
-import { ensureLoggedInAsync } from '../../../user/actions';
 import ProjectInit from '../init';
 
 jest.mock('fs');
 jest.mock('@expo/config');
-jest.mock('../../../project/projectUtils');
 jest.mock('../../../prompts');
 jest.mock('../../../user/actions');
 jest.mock('../../../graphql/mutations/AppMutation');
@@ -22,6 +22,7 @@ jest.mock('../../../ora', () => ({
   }),
 }));
 jest.mock('../../../project/fetchOrCreateProjectIDForWriteToConfigWithConfirmationAsync');
+jest.mock('../../../commandUtils/context/ProjectIdContextField');
 
 let originalProcessArgv: string[];
 
@@ -72,15 +73,14 @@ function mockTestProject(options: {
   const mockManifest = { exp: appJSON.expo };
 
   jest.mocked(getConfig).mockReturnValue(mockManifest as any);
-  jest.mocked(findProjectRootAsync).mockResolvedValue('/test-project');
+  jest.spyOn(ProjectDirContextField.prototype, 'getValueAsync').mockResolvedValue('/test-project');
+  jest.spyOn(ActorContextField.prototype, 'getValueAsync').mockResolvedValue(jester);
 }
 
 const commandOptions = { root: '/test-project' } as any;
 
 beforeEach(() => {
   jest.resetAllMocks();
-
-  jest.mocked(ensureLoggedInAsync).mockResolvedValue(jester);
 });
 
 describe(ProjectInit.name, () => {
@@ -93,20 +93,23 @@ describe(ProjectInit.name, () => {
       describe('interactive', () => {
         it('is no-op if already configured for id', async () => {
           await new ProjectInit(['--id', '1234'], commandOptions).run();
-          expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+          expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
         });
 
         it('prompts to overwrite when different', async () => {
           jest.mocked(confirmAsync).mockResolvedValue(true);
           await new ProjectInit(['--id', '12345'], commandOptions).run();
-          expect(saveProjectIdToAppConfigAsync).toHaveBeenCalledWith('/test-project', '12345');
+          expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).toHaveBeenCalledWith(
+            '/test-project',
+            '12345'
+          );
           expect(confirmAsync).toHaveBeenCalled();
         });
 
         it('aborts when prompt to overwrite is declined', async () => {
           jest.mocked(confirmAsync).mockResolvedValue(false);
           await new ProjectInit(['--id', '12345'], commandOptions).run();
-          expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+          expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
           expect(confirmAsync).toHaveBeenCalled();
         });
       });
@@ -116,7 +119,10 @@ describe(ProjectInit.name, () => {
           await new ProjectInit(['--id', '12345', '--force'], {
             root: '/test-project',
           } as any).run();
-          expect(saveProjectIdToAppConfigAsync).toHaveBeenCalledWith('/test-project', '12345');
+          expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).toHaveBeenCalledWith(
+            '/test-project',
+            '12345'
+          );
           expect(confirmAsync).not.toHaveBeenCalled();
         });
       });
@@ -132,7 +138,7 @@ describe(ProjectInit.name, () => {
               '1234'
             )}. Use --force flag to overwrite.`
           );
-          expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+          expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
         });
       });
     });
@@ -144,7 +150,10 @@ describe(ProjectInit.name, () => {
 
       it('configures', async () => {
         await new ProjectInit(['--id', '1234'], commandOptions).run();
-        expect(saveProjectIdToAppConfigAsync).toHaveBeenCalledWith('/test-project', '1234');
+        expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).toHaveBeenCalledWith(
+          '/test-project',
+          '1234'
+        );
       });
     });
   });
@@ -157,7 +166,7 @@ describe(ProjectInit.name, () => {
 
       it('does not configure', async () => {
         await new ProjectInit([], commandOptions).run();
-        expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+        expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
       });
     });
 
@@ -175,7 +184,10 @@ describe(ProjectInit.name, () => {
           it('prompts for confirmation to link', async () => {
             jest.mocked(confirmAsync).mockResolvedValue(true);
             await new ProjectInit([], commandOptions).run();
-            expect(saveProjectIdToAppConfigAsync).toHaveBeenCalledWith('/test-project', '123456');
+            expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).toHaveBeenCalledWith(
+              '/test-project',
+              '123456'
+            );
             expect(confirmAsync).toHaveBeenCalledTimes(1);
           });
 
@@ -184,7 +196,7 @@ describe(ProjectInit.name, () => {
             await expect(new ProjectInit([], commandOptions).run()).rejects.toThrow(
               'Project ID configuration canceled. Re-run the command to select a different account/project.'
             );
-            expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+            expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
             expect(confirmAsync).toHaveBeenCalledTimes(1);
           });
         });
@@ -194,7 +206,10 @@ describe(ProjectInit.name, () => {
             jest.mocked(confirmAsync).mockResolvedValue(true);
             jest.mocked(AppMutation.createAppAsync).mockResolvedValue('0129');
             await new ProjectInit([], commandOptions).run();
-            expect(saveProjectIdToAppConfigAsync).toHaveBeenCalledWith('/test-project', '0129');
+            expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).toHaveBeenCalledWith(
+              '/test-project',
+              '0129'
+            );
             expect(confirmAsync).toHaveBeenCalledTimes(1);
           });
 
@@ -203,7 +218,7 @@ describe(ProjectInit.name, () => {
             await expect(new ProjectInit([], commandOptions).run()).rejects.toThrowError(
               'Project ID configuration canceled for @jester/testing-123.'
             );
-            expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+            expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
             expect(confirmAsync).toHaveBeenCalledTimes(1);
           });
         });
@@ -229,7 +244,7 @@ describe(ProjectInit.name, () => {
         await expect(new ProjectInit([], commandOptions).run()).rejects.toThrowError(
           `You don't have permission to create a new project on the other account and no matching project already exists on the account.`
         );
-        expect(saveProjectIdToAppConfigAsync).not.toHaveBeenCalled();
+        expect(ProjectIdContextField['saveProjectIdToAppConfigAsync']).not.toHaveBeenCalled();
       });
     });
   });
