+++ packages/eas-cli/src/project/projectUtils.ts
@@ -1,22 +1,15 @@
-import { ExpoConfig, getConfigFilePaths, getPackageJson, modifyConfigAsync } from '@expo/config';
-import { Env } from '@expo/eas-build-job';
+import { ExpoConfig, getConfigFilePaths, getPackageJson } from '@expo/config';
 import chalk from 'chalk';
 import fs from 'fs-extra';
 import path from 'path';
-import pkgDir from 'pkg-dir';
 import resolveFrom from 'resolve-from';
 import semver from 'semver';
 
 import { AccountFragment, AppPrivacy } from '../graphql/generated';
 import { AppQuery } from '../graphql/queries/AppQuery';
 import Log from '../log';
-import { ora } from '../ora';
 import { Actor } from '../user/User';
-import { ensureLoggedInAsync } from '../user/actions';
 import { expoCommandAsync } from '../utils/expoCli';
-import { getVcsClient } from '../vcs';
-import { getExpoConfig } from './expoConfig';
-import { fetchOrCreateProjectIDForWriteToConfigWithConfirmationAsync } from './fetchOrCreateProjectIDForWriteToConfigWithConfirmationAsync';
 
 export function getUsername(exp: ExpoConfig, user: Actor): string | undefined {
   switch (user.__typename) {
@@ -32,143 +25,6 @@ export function getUsername(exp: ExpoConfig, user: Actor): string | undefined {
       // robot users don't have usernames
       return undefined;
   }
-}
-
-export async function findProjectRootAsync({
-  cwd,
-  defaultToProcessCwd = false,
-}: {
-  cwd?: string;
-  defaultToProcessCwd?: boolean;
-} = {}): Promise<string> {
-  const projectRootDir = await pkgDir(cwd);
-  if (!projectRootDir) {
-    if (!defaultToProcessCwd) {
-      throw new Error('Run this command inside a project directory.');
-    } else {
-      return process.cwd();
-    }
-  } else {
-    let vcsRoot;
-    try {
-      vcsRoot = path.normalize(await getVcsClient().getRootPathAsync());
-    } catch {}
-    if (vcsRoot && vcsRoot.startsWith(projectRootDir) && vcsRoot !== projectRootDir) {
-      throw new Error(
-        `package.json is outside of the current git repository (project root: ${projectRootDir}, git root: ${vcsRoot}.`
-      );
-    }
-    return projectRootDir;
-  }
-}
-
-/**
- * Save an EAS project ID to the appropriate field in the app config.
- */
-export async function saveProjectIdToAppConfigAsync(
-  projectDir: string,
-  projectId: string,
-  options: { env?: Env } = {}
-): Promise<void> {
-  const exp = getExpoConfig(projectDir, options);
-  const result = await modifyConfigAsync(projectDir, {
-    extra: { ...exp.extra, eas: { ...exp.extra?.eas, projectId } },
-  });
-
-  switch (result.type) {
-    case 'success':
-      break;
-    case 'warn': {
-      Log.warn();
-      Log.warn(
-        `Warning: Your project uses dynamic app configuration, and the EAS project ID can't automatically be added to it.`
-      );
-      Log.warn(
-        chalk.dim(
-          'https://docs.expo.dev/workflow/configuration/#dynamic-configuration-with-appconfigjs'
-        )
-      );
-      Log.warn();
-      Log.warn(
-        `To complete the setup process, set the ${chalk.bold(
-          'extra.eas.projectId'
-        )} in your ${chalk.bold(getProjectConfigDescription(projectDir))}:`
-      );
-      Log.warn();
-      Log.warn(chalk.bold(JSON.stringify({ expo: { extra: { eas: { projectId } } } }, null, 2)));
-      Log.warn();
-      throw new Error(result.message);
-    }
-    case 'fail':
-      throw new Error(result.message);
-    default:
-      throw new Error('Unexpected result type from modifyConfigAsync');
-  }
-}
-
-/**
- * Get the EAS project ID from the app config. If the project ID is not set in the config,
- * use the owner/slug to identify an EAS project on the server (asking for confirmation first),
- * and attempt to save the EAS project ID to the appropriate field in the app config. If unable to
- * save to the app config, throw an error.
- */
-export async function getProjectIdAsync(
-  exp: ExpoConfig,
-  options: { env?: Env; nonInteractive: boolean },
-  findProjectRootOptions: {
-    cwd?: string;
-    defaultToProcessCwd?: boolean;
-  } = {}
-): Promise<string> {
-  const localProjectId = exp.extra?.eas?.projectId;
-  if (localProjectId) {
-    return localProjectId;
-  }
-
-  const projectDir = await findProjectRootAsync(findProjectRootOptions);
-  if (!projectDir) {
-    throw new Error('This command must be run inside a project directory.');
-  }
-
-  Log.warn('EAS project not configured.');
-
-  const getAccountNameForEASProjectSync = (exp: ExpoConfig, user: Actor): string => {
-    if (exp.owner) {
-      return exp.owner;
-    }
-    switch (user.__typename) {
-      case 'User':
-        return user.username;
-      case 'Robot':
-        throw new Error(
-          'The "owner" manifest property is required when using robot users. See: https://docs.expo.dev/versions/latest/config/app/#owner'
-        );
-    }
-  };
-
-  const actor = await ensureLoggedInAsync({ nonInteractive: options.nonInteractive });
-  const projectId = await fetchOrCreateProjectIDForWriteToConfigWithConfirmationAsync(
-    {
-      accountName: getAccountNameForEASProjectSync(exp, actor),
-      projectName: exp.slug,
-      privacy: toAppPrivacy(exp.privacy),
-    },
-    {
-      nonInteractive: options.nonInteractive,
-    },
-    actor
-  );
-
-  const spinner = ora(`Linking local project to EAS project ${projectId}`).start();
-  try {
-    await saveProjectIdToAppConfigAsync(projectDir, projectId, options);
-    spinner.succeed(`Linked local project to EAS project ${projectId}`);
-  } catch (e: any) {
-    spinner.fail();
-    throw e;
-  }
-
-  return projectId;
 }
 
 export const toAppPrivacy = (privacy: ExpoConfig['privacy']): AppPrivacy => {
